- 前端修改 ：
- 移除了 LeaveRequest.vue 中的开始时间和结束时间选择器
- 修改了 handleShiftChange 函数，增加了天数计算逻辑
- 添加了 watch 监听器，监听日期变化并自动更新请假时长
- 确保了与后端 API 的兼容性，保留了 startTime 和 endTime 字段
- 计算逻辑 ：

- 班次总工时：所有选中班次的工时之和
- 请假天数：结束日期 - 开始日期 + 1（包含起止日期）
- 总请假时长：班次总工时 × 请假天数

+++ src/main/java/com/lasu/hyperduty/service/impl/DutyStatisticsServiceImpl.java
@@ -378,7 +378,7 @@ public class DutyStatisticsServiceImpl extends ServiceImpl<DutyStatisticsMapper,
                     request.getLeaveType() != null && 
                     request.getLeaveType() == 4 && // 调休类型
                     request.getTotalHours() != null) {
-                    compensatoryHours = compensatoryHours.subtract(BigDecimal.valueOf(request.getTotalHours()));
+                    compensatoryHours = compensatoryHours.subtract(request.getTotalHours());
                 }
             }
